// src/problem2/src/types/enums.ts
export enum EActiveInput {
  FROM = "from",
  TO = "to",
}

export enum ESwapResult {
  SUCCESS = "success",
  FAILED = "failed",
}

export enum EButtonVariant {
  PRIMARY = "primary",
  SECONDARY = "secondary",
  TONER = "toner",
}

export enum ETokenSize {
  SM = "sm",
  MD = "md",
  LG = "lg",
}